requirements.txt
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.8
deepdiff==6.7.1
dipy==1.8.0
display==1.0.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
h5py==3.10.0
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
nibabel==5.2.0
numpy==1.24.4
ordered-set==4.1.0
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.10.1
setuptools-scm==8.0.4
six==1.16.0
sympy==1.12
TIME-python==0.0.17
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
trx-python==0.2.9
typing_extensions==4.9.0
zipp==3.17.0
